requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
django-jazzmin==2.6.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
PyJWT==1.7.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
